backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
ffmpeg-python==0.2.0
pytesseract==0.3.10
imagehash==4.3.1
numpy==1.26.3
pydantic==2.5.3
aiofiles==23.2.1
python-dotenv==1.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0
python-jose[cryptography]==3.3.0
pyjwt==2.8.0
httpx==0.26.0
